[ADD] hitbox, colision system, movement system, direction

+++ engine/includes/system/DisplaySystem.hpp
@@ -20,7 +20,7 @@ public:
         std::vector<std::shared_ptr<TEntity>> filteredEntities;
         for (const std::shared_ptr<TEntity>& entity : _entities) {
             if (entity->hasComponent(typeid(Sprite<std::shared_ptr<sf::Sprite>>)) &&
-                entity->hasComponent(typeid(Position<std::pair<int, int>>))) {
+                entity->hasComponent(typeid(Position<std::pair<double, double>>))) {
                 filteredEntities.push_back(entity);
             }
         }
@@ -40,7 +40,7 @@ public:
     void execute(std::vector<std::shared_ptr<TEntity>>& _entities, sf::RenderWindow &_window, std::vector<int> _inputs) override {
         for (const std::shared_ptr<TEntity>& entity : _entities) {
             std::shared_ptr<sf::Sprite> spriteComp = entity->template getComponent<Sprite<std::shared_ptr<sf::Sprite>>>()->getValue();
-            std::shared_ptr<Position<std::pair<int, int>>> positionComp = entity->template getComponent<Position<std::pair<int, int>>>();
+            std::shared_ptr<Position<std::pair<double, double>>> positionComp = entity->template getComponent<Position<std::pair<double, double>>>();
             float xPosition = static_cast<float>(positionComp->getValue().first);
             float yPosition = static_cast<float>(positionComp->getValue().second);
